+++ src/client/renderer/OffsettedRenderList.h
@@ -9,7 +9,7 @@ class OffsettedRenderList
 private:
 	int_t x = 0, y = 0, z = 0;
 	std::vector<int_t> lists; // = MemoryTracker::createIntBuffer(0x10000);
-	float xOff = 0.0f, yOff = 0.0f, zOff = 0.0f;
+	double xOff = 0.0, yOff = 0.0, zOff = 0.0;  // Store as double to preserve precision at long distances
 	bool inited = false;
 	bool rendered = false;
